Forgot to push fix to issue 7. This should fix issue 7.
As for the other issue, SimpleMapIO has a boolean option now for whether the user wants to open or write to a file. This should prevent the file from being erased when opened. The previous String method compared pointers for Strings instead of the Strings themselves, and would have had unexpected behavior. Hopefully this fixes the MapIO issue. Also I added Mark as an author to the two classes he has contributed to.